Prep Work
Due to issues with the local machines in the university not having JDK 8 8u301. I cannot launch games on debugged android devices. This means that I cannot test the games on my phone without an external system. I will begin bringing my laptop to the university again to resolve this in future cases.